Refactor Metal.NET generator: fix type mappings, enum handling, struct returns, and code style

+++ Metal.NET/Metal/MTL4BinaryFunctionOptions.cs
@@ -1,7 +1,7 @@
 ﻿namespace Metal.NET;
 
 [Flags]
-public enum MTL4BinaryFunctionOptions : uint
+public enum MTL4BinaryFunctionOptions : ulong
 {
     BinaryFunctionOptionNone = 0,
